+++ lib/plugins/aws/package/compile/functions.js
