+++ lib/plugins/index.js
@@ -33,7 +33,7 @@ module.exports = [
   require('./aws/rollback/index.js'),
   require('./aws/rollbackFunction/index.js'),
   require('./aws/package/compile/layers/index.js'),
-  require('./aws/package/compile/functions/index.js'),
+  require('./aws/package/compile/functions.js'),
   require('./aws/package/compile/events/schedule.js'),
   require('./aws/package/compile/events/s3/index.js'),
   require('./aws/package/compile/events/apiGateway/index.js'),
